+++ src/CascadeMultiSelect.jsx
@@ -10,10 +10,13 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import Dropdown from 'uxcore-dropdown';
 import Button from 'uxcore-button';
+import { polyfill } from 'react-lifecycles-compat';
 import CascadeMultiPanel from './CascadeMultiPanel';
 import CascadeMultiModal from './CascadeMultiModal';
 import i18n from './locale';
 import { getDisabledValueLabel, getWidthStyle } from './utils';
+
+const CASCADER_SELECT_PREFIXCLS = 'kuma-cascader';
 
 const makeOptionsChecked = (value = [], options) => {
   // 没有value则需要设置check为false
@@ -34,30 +37,57 @@ const makeOptionsChecked = (value = [], options) => {
 };
 
 class CascadeMultiSelect extends React.Component {
+  static separator = ' , ';
+
+  static getDerivedStateFromProps(props, state) {
+    const { value } = props;
+    const displayValue = CascadeMultiSelect.getInputValue(props, value);
+    return { displayValue, value };
+  }
+
+  static getInputValue(props, value) {
+    const { options, beforeRender, locale } = props;
+    if (beforeRender) {
+      makeOptionsChecked(value, options);
+      return beforeRender(value, options);
+    }
+
+    const arr = [];
+    if (value && value.length) {
+      for (let i = 0; i < value.length; i += 1) {
+        arr.push(CascadeMultiSelect.getValueLabel(options, value[i], locale));
+      }
+    }
+    return arr.join(CascadeMultiSelect.separator);
+  }
+
+  static getValueLabel(dataList, key, locale) {
+    let back = '';
+    if (dataList && dataList.length) {
+      for (let i = 0; i < dataList.length; i += 1) {
+        if (dataList[i].value === key) {
+          return dataList[i].children && dataList[i].children.length ?
+            `${dataList[i].label} (${i18n(locale).all})` :
+            dataList[i].label;
+        }
+        if (dataList[i].children) {
+          const res = CascadeMultiSelect.getValueLabel(dataList[i].children, key, locale);
+          back = res || back;
+        }
+      }
+    }
+    return back;
+  }
 
   constructor(props) {
     super(props);
     this.state = {
-      value: props.value,
-      defaultValue: props.defaultValue,
+      value: props.value || props.defaultValue,
       displayValue: '',
-      allowClear: props.allowClear,
-      disabled: props.disabled,
       showSubMenu: false,
       result: {},
     };
-    this.separator = ' , ';
-    this.data = {
-      value: props.value || props.defaultValue,
-      options: props.options,
-      displayValue: '',
-      result: {},
-      size: props.size,
-    };
     this.hasChanged = false;
-  }
-
-  componentWillMount() {
     this.onOk = this.onOk.bind(this);
     this.handleSelect = this.handleSelect.bind(this);
     this.handleItemClick = this.handleItemClick.bind(this);
@@ -65,20 +95,8 @@ class CascadeMultiSelect extends React.Component {
   }
 
   componentDidMount() {
-    const { value, defaultValue, options } = this.props;
-    const displayValue = this.getInputValue(value, options);
-    this.data.displayValue = displayValue;
-    this.data.value = value;
-    this.setInputValue(value || defaultValue, options);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    // bugfix: 当 props.value 首先传递进组件之后再传递 options 数据并没有回填
-    const { value, options } = nextProps;
-    const displayValue = this.getInputValue(value, options);
-    this.data.displayValue = displayValue;
-    this.data.value = value;
-    this.setInputValue(value, options);
+    const { value, defaultValue } = this.props;
+    this.setInputValue(value || defaultValue);
   }
 
   onOk() {
@@ -86,11 +104,8 @@ class CascadeMultiSelect extends React.Component {
       return;
     }
     const { value, result } = this.state;
-    const displayValue = this.getInputValue(value, this.props.options);
+    const displayValue = CascadeMultiSelect.getInputValue(this.props, value);
     const { valueList, labelList, leafList, cascadeSelected } = result;
-    this.data.displayValue = displayValue;
-    this.data.value = value;
-    this.data.result = result;
     this.setState({
       displayValue,
       value,
@@ -100,9 +115,9 @@ class CascadeMultiSelect extends React.Component {
   }
 
   onCancel() {
-    const { value, options, result } = this.data;
+    const { value, result } = this.state;
     this.setState({
-      displayValue: this.getInputValue(value, options),
+      displayValue: CascadeMultiSelect.getInputValue(this.props, value),
       value,
       result,
     }, () => {
@@ -128,10 +143,8 @@ class CascadeMultiSelect extends React.Component {
       result.labelList = labelList;
       result.valueList = value;
       result.leafList = leafList;
-      displayValue = this.getInputValue(value, options);
+      displayValue = CascadeMultiSelect.getInputValue(this.props, value);
     }
-    this.data.displayValue = displayValue;
-    this.data.value = value;
     this.setState({
       value,
       displayValue,
@@ -153,21 +166,6 @@ class CascadeMultiSelect extends React.Component {
     }
   }
 
-  getInputValue(value, dataList) {
-    if (this.props.beforeRender) {
-      makeOptionsChecked(value, dataList);
-      return this.props.beforeRender(value, dataList);
-    }
-
-    const arr = [];
-    if (value && value.length) {
-      for (let i = 0; i < value.length; i += 1) {
-        arr.push(this.getValueLabel(dataList, value[i]));
-      }
-    }
-    return arr.join(this.separator);
-  }
-
   getValueLabel(dataList, key) {
     let back = '';
     if (dataList && dataList.length) {
@@ -186,8 +184,8 @@ class CascadeMultiSelect extends React.Component {
     return back;
   }
 
-  setInputValue(value, dataList) {
-    const displayValue = this.getInputValue(value, dataList);
+  setInputValue(value) {
+    const displayValue = CascadeMultiSelect.getInputValue(this.props, value);
     this.setState({ displayValue, value });
   }
 
@@ -206,9 +204,8 @@ class CascadeMultiSelect extends React.Component {
   }
 
   handleSelect(valueList, labelList, leafList, cascadeSelected) {
-    const { options } = this.props;
     this.setState({
-      displayValue: this.getInputValue(valueList, options),
+      displayValue: CascadeMultiSelect.getInputValue(this.props),
       value: valueList,
       result: {
         valueList,
@@ -234,9 +231,8 @@ class CascadeMultiSelect extends React.Component {
   }
 
   renderInput() {
-    const { prefixCls, placeholder, locale, readOnly } = this.props;
-    const { disabled } = this.state;
-    const { displayValue } = this.data;
+    const { prefixCls, placeholder, locale, readOnly, disabled } = this.props;
+    const { displayValue } = this.state;
     if (readOnly) {
       return <span>{displayValue}</span>;
     }
@@ -244,7 +240,7 @@ class CascadeMultiSelect extends React.Component {
       <div>
         {
           !displayValue.length ?
-            <div className="kuma-cascader-placeholder">
+            <div className={`${CASCADER_SELECT_PREFIXCLS}-placeholder`}>
               {placeholder || i18n(locale).placeholder}
             </div> :
             <div className={classnames([`${prefixCls}-text-result`])}>
@@ -263,10 +259,10 @@ class CascadeMultiSelect extends React.Component {
   }
 
   renderCloseIcon() {
-    const { disabled } = this.state;
+    const { disabled } = this.props;
     if (disabled) { return null; }
     return (
-      <div className={'kuma-cascader-close-wrap'}>
+      <div className={`${CASCADER_SELECT_PREFIXCLS}-close-wrap`}>
         <i
           className="kuma-icon kuma-icon-error"
           onClick={(e) => {
@@ -280,34 +276,34 @@ class CascadeMultiSelect extends React.Component {
   }
 
   renderContent() {
-    const { className, prefixCls, size } = this.props;
-    const { displayValue, allowClear, disabled, showSubMenu } = this.state;
-    const prefixCls2 = 'kuma-cascader';
+    const { className, prefixCls, size, allowClear, disabled } = this.props;
+    const { displayValue, showSubMenu } = this.state;
+    
     return (
       <div
         className={classnames({
           [className]: true,
           [`${prefixCls}-input`]: !disabled,
           [`${prefixCls}-${size}`]: !!size,
-          [`${prefixCls2}-wrapper`]: true,
-          [`${prefixCls2}-disabled`]: disabled,
-          [`${prefixCls2}-clearable`]: !disabled && allowClear && displayValue.length > 0,
+          [`${CASCADER_SELECT_PREFIXCLS}-wrapper`]: true,
+          [`${CASCADER_SELECT_PREFIXCLS}-disabled`]: disabled,
+          [`${CASCADER_SELECT_PREFIXCLS}-clearable`]: !disabled && allowClear && displayValue.length > 0,
         })}
       >
         <div
           className={classnames({
-            [`${prefixCls2}-text`]: true,
+            [`${CASCADER_SELECT_PREFIXCLS}-text`]: true,
             [`${prefixCls}-${size}`]: !!size,
           })}
         >
-          <div className={`${prefixCls2}-trigger`}>
+          <div className={`${CASCADER_SELECT_PREFIXCLS}-trigger`}>
             {this.renderInput()}
           </div>
         </div>
         <div
           className={classnames({
-            [`${prefixCls2}-arrow`]: true,
-            [`${prefixCls2}-arrow-reverse`]: showSubMenu,
+            [`${CASCADER_SELECT_PREFIXCLS}-arrow`]: true,
+            [`${CASCADER_SELECT_PREFIXCLS}-arrow-reverse`]: showSubMenu,
           })}
         >
           <i className="kuma-icon kuma-icon-triangle-down" />
@@ -440,5 +436,7 @@ CascadeMultiSelect.displayName = 'CascadeMultiSelect';
 
 CascadeMultiSelect.CascadeMultiPanel = CascadeMultiPanel;
 CascadeMultiSelect.CascadeMultiModal = CascadeMultiModal;
+
+polyfill(CascadeMultiSelect);
 
 module.exports = CascadeMultiSelect;